+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 **1.6.1 TBD
 * Router - avoid regex matching when possible #442
 * Make `@JsonView` annotation an alias of `@ResponseContentType(H.Media.JSON)` #440
+* Support Content-Security-Policy header #439
 * Support using MACRO for URL path regex definition #438
 
 **1.6.0 28/Dec/2017**